Add simple cherrypy server to interface via sms
Adds a simple cherrypy server setup that will allow the game to be played
via texting. As of now the server only receives a message and responds with
a constant reply, but once more framework is set up it will be easy to get
the replies from the game engine itself.