Add command to create games
This creates a basic game manager which parses commands sent from the user.
The commands can either be meant to manage games (such as creating a game)
or the commands may be gameplay commands for specific games. This commit
adds an option to the parser to allow for game creation.