Fix behavioural bug in shuffling
Previously shuffling would create a new deck, implying that all the
cards had been "picked up" and were being shuffled for a new hand.
However, this did not pick up the visible cards from the game, so the
visible cards would just continue to grow unchecked. This clears the
visible cards when shuffling occurs so it truly is a new deck.